+++ providers/google/src/main/scala/com/colisweb/distances/providers/google/GoogleGeoApiContext.scala
@@ -2,12 +2,20 @@ package com.colisweb.distances.providers.google
 
 import java.util.concurrent.TimeUnit
 
-import com.google.maps.GeoApiContext
+import com.google.maps.{GeoApiContext, OkHttpRequestHandler}
+import okhttp3.logging.HttpLoggingInterceptor
+import okhttp3.logging.HttpLoggingInterceptor.Level
 
 import scala.concurrent.duration._
 import scala.language.postfixOps
 
-final case class GoogleGeoApiContext(apiKey: String, connectTimeout: Duration, readTimeout: Duration, queryRateLimit: Int) {
+final case class GoogleGeoApiContext(
+    apiKey: String,
+    connectTimeout: Duration,
+    readTimeout: Duration,
+    queryRateLimit: Int,
+    loggingF: String => Unit
+) {
   assert(apiKey.trim.nonEmpty, "apiKey must be a non empty String")
 
   /**
@@ -27,16 +35,22 @@ final case class GoogleGeoApiContext(apiKey: String, connectTimeout: Duration, r
     *  - 1000 elements per second (EPS), calculated as the sum of client-side and server-side queries.
     * ```
     */
+  private val logging: HttpLoggingInterceptor = new HttpLoggingInterceptor((message: String) => loggingF(message))
+  logging.level(Level.BASIC)
+
+  private val builder = new OkHttpRequestHandler.Builder()
+  builder.okHttpClientBuilder().addInterceptor(logging)
+
   val geoApiContext: GeoApiContext =
-    new GeoApiContext.Builder()
+    new GeoApiContext.Builder(builder)
       .apiKey(apiKey)
       .connectTimeout(connectTimeout.toMillis, TimeUnit.MILLISECONDS)
       .readTimeout(readTimeout.toMillis, TimeUnit.MILLISECONDS)
       .queryRateLimit(queryRateLimit)
       .build()
-
 }
 
 object GoogleGeoApiContext {
-  final def apply(googleApiKey: String): GoogleGeoApiContext = new GoogleGeoApiContext(googleApiKey, 1 second, 1 second, 1000)
+  final def apply(googleApiKey: String, loggingF: String => Unit = _ => ()): GoogleGeoApiContext =
+    new GoogleGeoApiContext(googleApiKey, 1 second, 1 second, 1000, loggingF)
 }