+++ tests/src/test/scala/com/colisweb/distances/generators/Gens.scala
@@ -1,5 +1,6 @@
 package com.colisweb.distances.generators
 
+import com.colisweb.distances.TravelMode
 import com.colisweb.distances.Types._
 import com.colisweb.distances.caches.Toto
 import org.scalacheck.{Arbitrary, Gen}